// src/types/Routes.ts
export enum Routes {
  App = "/app",
  Dashboard = "/app/dashboard",
  Wallet = "/app/wallet",
  Transactions = "/app/transactions",
  Account = "/app/account",
  Auth = "/auth",
  Login = "/auth/login",
  Register = "/auth/register",
  Home = "/",
  Demo = "/demo",
  DashboardDemo = "/demo/dashboard",
  WalletDemo = "/demo/wallet",
  TransactionsDemo = "/demo/transactions",
  AccountDemo = "/demo/account",
  ForgotPassword = "/auth/forgot-password",
  About = "/about",
}